Make ControlledListBox the ListBox
